numbers
I have written a smart contract in Solidity that allows users to perform basic mathematical operations on two input numbers, "num1" and "num2". The contract includes functions for addition, subtraction, multiplication, and division. When the "numbers" function is called with the input values for "num1" and "num2", the contract performs the calculations and stores the results in internal variables.

For example, if you want to get the sum of "num1" and "num2", you can call the "add" function, which will return the sum. Similarly, you can call the "subtract", "multiply", and "divide" functions to get the results of the respective mathematical operations.